+++ packages/uno-serverless-aws/test/unit/services/key-value-repository-test.ts
@@ -1,7 +1,6 @@
 import { AWSError, Response, S3 } from "aws-sdk";
 import { expect } from "chai";
-import * as HttpStatusCodes from "http-status-codes";
-import { HealthCheckStatus, randomStr } from "uno-serverless";
+import { HealthCheckStatus, HttpStatusCodes, randomStr } from "uno-serverless";
 import { S3KeyValueRepository } from "../../../src/services/key-value-repository";
 import { S3Client } from "../../../src/services/s3-client";