+++ packages/uno-serverless-azure/webpack.config.js
@@ -32,7 +32,6 @@ module.exports = {
   externals: {
     'uno-serverless': 'uno-serverless',
     'aws-sdk': 'aws-sdk',
-    'http-status-codes': 'http-status-codes',
     'documentdb': 'documentdb',
     'azure-keyvault': 'azure-keyvault',
     'azure-storage': 'azure-storage',