+++ packages/uno-serverless/src/core/http-status-codes.ts
@@ -1,0 +1,2 @@
+import * as StatusCodes from "http-status-codes";
+export const HttpStatusCodes = StatusCodes;