+++ packages/uno-serverless/src/core/index.ts
@@ -1,6 +1,7 @@
 export * from "./container";
 export * from "./continuation";
 export * from "./errors";
+export * from "./http-status-codes";
 export * from "./json-schema";
 export * from "./schemas";
 export * from "./uno";