+++ packages/uno-serverless/test/unit/middlewares/http-test.ts
@@ -1,7 +1,7 @@
 import { expect } from "chai";
-import * as HttpStatusCodes from "http-status-codes";
 import { describe, it } from "mocha";
 import { notFoundError } from "../../../src/core/errors";
+import { HttpStatusCodes } from "../../../src/core/http-status-codes";
 import { HttpUnoEvent } from "../../../src/core/schemas";
 import { testAdapter, uno } from "../../../src/core/uno";
 import { randomStr } from "../../../src/core/utils";